more vs 2017 warning fixes

+++ kernels/builders/heuristic_binning.h
@@ -241,7 +241,7 @@ namespace embree
           const unsigned int b00 = extract<0>(bin0); bounds(b00,0).extend(prim0); 
           const unsigned int b01 = extract<1>(bin0); bounds(b01,1).extend(prim0); 
           const unsigned int b02 = extract<2>(bin0); bounds(b02,2).extend(prim0); 
-          const unsigned int s0 = prims[i+0].size();
+          const unsigned int s0 = (unsigned int)prims[i+0].size();
           counts(b00,0)+=s0;
           counts(b01,1)+=s0;
           counts(b02,2)+=s0;
@@ -250,7 +250,7 @@ namespace embree
           const unsigned int b10 = extract<0>(bin1);  bounds(b10,0).extend(prim1); 
           const unsigned int b11 = extract<1>(bin1);  bounds(b11,1).extend(prim1); 
           const unsigned int b12 = extract<2>(bin1);  bounds(b12,2).extend(prim1); 
-          const unsigned int s1 = prims[i+1].size();
+          const unsigned int s1 = (unsigned int)prims[i+1].size();
           counts(b10,0)+=s1;
           counts(b11,1)+=s1;
           counts(b12,2)+=s1;
@@ -264,7 +264,7 @@ namespace embree
           const vint4 bin0 = (vint4)mapping.bin(center0); 
           
           /*! increase bounds of bins */
-          const unsigned int s0 = prims[i].size();
+          const unsigned int s0 = (unsigned int)prims[i].size();
           const int b00 = extract<0>(bin0); counts(b00,0)+=s0; bounds(b00,0).extend(prim0);
           const int b01 = extract<1>(bin0); counts(b01,1)+=s0; bounds(b01,1).extend(prim0);
           const int b02 = extract<2>(bin0); counts(b02,2)+=s0; bounds(b02,2).extend(prim0);
